login work in progress hash password done

+++ src/Controller/UserController.php
@@ -9,10 +9,19 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
+use Symfony\Component\PasswordHasher\Hasher\UserPasswordHasherInterface;
 
 #[Route('/user')]
 class UserController extends AbstractController
 {
+    private $passwordHasher;
+
+    public function __construct(UserPasswordHasherInterface $passwordHasher)
+    {
+        $this->passwordHasher = $passwordHasher;
+    }
+
     #[Route('/', name: 'app_user_index', methods: ['GET'])]
     public function index(UserRepository $userRepository): Response
     {
@@ -27,17 +36,29 @@ class UserController extends AbstractController
         $user = new User();
         $form = $this->createForm(UserType::class, $user);
         $form->handleRequest($request);
+    
         if ($form->isSubmitted() && $form->isValid()) {
+            $plainPassword = $form->get('password')->getData(); // Remplacez 'password' par le nom du champ du formulaire
+    
+            $hashedPassword = $this->passwordHasher->hashPassword(
+                $user,
+                $plainPassword
+            );
+            $user->setPassword($hashedPassword);
+    
             $userRepository->save($user, true);
+    
             return $this->redirectToRoute('app_photo_index', [], Response::HTTP_SEE_OTHER);
         }
+    
         return $this->renderForm('user/new.html.twig', [
             'user' => $user,
             'form' => $form,
         ]);
     }
+    
 
-    #[Route('/{id}', name: 'app_user_show', methods: ['GET'])]
+    #[Route('/{id}/show', name: 'app_user_show', methods: ['GET'])]
     public function show(User $user): Response
     {
         return $this->render('user/show.html.twig', [
@@ -63,13 +84,20 @@ class UserController extends AbstractController
         ]);
     }
 
-    #[Route('/{id}', name: 'app_user_delete', methods: ['POST'])]
-    public function delete(Request $request, User $user, UserRepository $userRepository): Response
+    #[Route('/{id}/delete', name: 'app_user_delete', methods: ['POST'])]
+    public function delete(Request $request, User $user, UserRepository $userRepository, UserPasswordHasherInterface $passwordHasher): Response
     {
         if ($this->isCsrfTokenValid('delete'.$user->getId(), $request->request->get('_token'))) {
+            $confirmPassword = $request->request->get('confirm_password');
+    
+            if (!$passwordHasher->isPasswordValid($user, $confirmPassword)) {
+                throw new AccessDeniedHttpException();
+            }
+    
             $userRepository->remove($user, true);
         }
-
+    
         return $this->redirectToRoute('app_user_index', [], Response::HTTP_SEE_OTHER);
     }
+    
 }
